+++ database/migrations/2024_06_06_110035_create_calendar.php
@@ -11,24 +11,26 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('calendar', function (Blueprint $table) {
-            $table->id('idCalendar');
-            $table->unsignedBigInteger('Prix_idPrix');
-            $table->unsignedBigInteger('GrandPrix_idGrandPrix');
-            $table->dateTime('t1');
-            $table->dateTime('t2')->nullable();
-            $table->dateTime('t3')->nullable();
-            $table->dateTime('sprintQualify')->nullable();
-            $table->dateTime('sprint')->nullable();
-            $table->dateTime('qualify');
-            $table->dateTime('race');
-            
-            $table->unique('idCalendar');
-            $table->foreign('Prix_idPrix')->references('idPrix')->on('prix')->onDelete('cascade');
-            $table->foreign('GrandPrix_idGrandPrix')->references('idGrandPrix')->on('grandprix')->onDelete('cascade');
+        if (!Schema::hasTable('calendar')) {
+            Schema::create('calendar', function (Blueprint $table) {
+                $table->id('idCalendar');
+                $table->unsignedBigInteger('Prix_idPrix');
+                $table->unsignedBigInteger('GrandPrix_idGrandPrix');
+                $table->dateTime('t1');
+                $table->dateTime('t2')->nullable();
+                $table->dateTime('t3')->nullable();
+                $table->dateTime('sprintQualify')->nullable();
+                $table->dateTime('sprint')->nullable();
+                $table->dateTime('qualify');
+                $table->dateTime('race');
 
-            $table->timestamps(); // Se você desejar incluir os campos created_at e updated_at
-        });
+                $table->unique('idCalendar');
+                $table->foreign('Prix_idPrix')->references('idPrix')->on('prix')->onDelete('cascade');
+                $table->foreign('GrandPrix_idGrandPrix')->references('idGrandPrix')->on('grandprix')->onDelete('cascade');
+
+                $table->timestamps(); // Se você desejar incluir os campos created_at e updated_at
+            });
+        }
     }
 
     /**
